get cookies automatically
